requirements.txt
boto3
py-crunchbase-api
botocore
setuptools